Reorganize test files next to source files and export comparatorMatches
Move test files from __tests__ directories to sit alongside their source files.
Export comparatorMatches from stage-helpers.ts instead of duplicating in tests.
Update world-operator.ts to import comparatorMatches from stage-helpers.
